+++ packages/adp-tooling/test/unit/base/helper.test.ts
@@ -289,7 +289,7 @@ describe('helper', () => {
             });
         });
 
-        test('should filter out inbounds with hideLauncher not false', () => {
+        test('should filter out inbounds with hideLauncher equal to true', () => {
             const inbounds = [
                 {
                     content: {
@@ -307,8 +307,7 @@ describe('helper', () => {
                         action: 'action2',
                         title: 'Test Action 2',
                         description: 'Description 2',
-                        url: '/test/action2',
-                        hideLauncher: false
+                        url: '/test/action2'
                     }
                 }
             ] as unknown as Inbound[];
@@ -321,8 +320,7 @@ describe('helper', () => {
                     action: 'action2',
                     title: 'Test Action 2',
                     description: 'Description 2',
-                    url: '/test/action2',
-                    hideLauncher: false
+                    url: '/test/action2'
                 }
             });
         });
@@ -331,60 +329,6 @@ describe('helper', () => {
             const result = filterAndMapInboundsToManifest([]);
 
             expect(result).toBeUndefined();
-        });
-
-        test('should filter out parameters with invalid entries', () => {
-            const inbounds = [
-                {
-                    content: {
-                        semanticObject: 'Test',
-                        action: 'action1',
-                        signature: {
-                            parameters: {
-                                param1: {
-                                    defaultValue: { format: '', value: '' },
-                                    filter: { format: '' },
-                                    launcherValue: { value: 'test' }
-                                },
-                                param2: {
-                                    defaultValue: { format: 'plain', value: 'value' },
-                                    filter: { format: '' },
-                                    launcherValue: { value: 'test' }
-                                },
-                                param3: {
-                                    defaultValue: { format: 'plain', value: 'value' },
-                                    filter: { format: 'plain' },
-                                    launcherValue: { value: 'test', additionalProp: 'extra' }
-                                }
-                            }
-                        },
-                        hideLauncher: false
-                    }
-                }
-            ] as unknown as Inbound[];
-
-            const result = filterAndMapInboundsToManifest(inbounds);
-
-            expect(result).toEqual({
-                'Test-action1': {
-                    semanticObject: 'Test',
-                    action: 'action1',
-                    signature: {
-                        parameters: {
-                            param2: {
-                                defaultValue: { format: 'plain', value: 'value' },
-                                launcherValue: { value: 'test' }
-                            },
-                            param3: {
-                                defaultValue: { format: 'plain', value: 'value' },
-                                filter: { format: 'plain' },
-                                launcherValue: { value: 'test' }
-                            }
-                        }
-                    },
-                    hideLauncher: false
-                }
-            });
         });
     });
 });