+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/app-config-writer
 
+## 0.6.25
+
+### Patch Changes
+
+-   Updated dependencies [4e0bd83]
+    -   @sap-ux/project-access@1.30.8
+    -   @sap-ux/axios-extension@1.22.4
+
+## 0.6.24
+
+### Patch Changes
+
+-   Updated dependencies [ffac61c]
+    -   @sap-ux/axios-extension@1.22.4
+
 ## 0.6.23
 
 ### Patch Changes