+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.5.0
+
+### Minor Changes
+
+-   6e2c5aa: fix overwrite prompt conflict
+
+## 0.4.1
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.28
+
+## 0.4.0
+
+### Minor Changes
+
+-   089b56f: updates extension types and extends prompt options in sub gens
+
+## 0.3.30
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.27
+
 ## 0.3.29
 
 ### Patch Changes