+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.1.23
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.13.4
+
+## 0.1.22
+
+### Patch Changes
+
+-   Updated dependencies [ca44076]
+    -   @sap-ux/fiori-generator-shared@0.13.3
+
 ## 0.1.21
 
 ### Patch Changes