+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.6.11
+
+### Patch Changes
+
+-   Updated dependencies [4e0bd83]
+    -   @sap-ux/project-access@1.30.8
+    -   @sap-ux/cds-odata-annotation-converter@0.6.4
+
+## 0.6.10
+
+### Patch Changes
+
+-   2cae662: fix(deps): update dependency @sap-ux/annotation-converter to v0.10.3
+
 ## 0.6.9
 
 ### Patch Changes