+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.4.31
+
+### Patch Changes
+
+-   @sap-ux/project-input-validator@0.6.13
+-   @sap-ux/inquirer-common@0.7.28
+
+## 0.4.30
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.27
+
 ## 0.4.29
 
 ### Patch Changes